move make_physics_creature
